Major changes to ensure that W and Z are unambiguous
This major change forces the W and Z lepton association to be
unambigious. Note that in the process it brakes ZZ (to fix later...)

As it stood, the W,Z, and lepton collections were made separately.
Without truth information in the gen particles, this caused serious
problems for disambiguation, as the W/Z collections could be formed from
the same leptons or others besides the 3 leading in pt. This change
ignores the lepton collection passed to the Ntuplizer and instead uses
those in the W/Z. The first Z is taken (closest to Z mass).

The first W found which does not overlap the Z is then taken. It is
possible that this would not be the 3rd highest pt lepton -- though
perhaps the W candidates are in the order of the lepton candidates,
which are pt sorted.